service_type 7 && monitor all states !0
Service type 7 (program) collects the output of the service in a list of xml elements. Check if this list is empty, concatenate all list values or set 'no command output available' as message for non existing output.

Instead of only taking in account monitor_states 1 (Yes) and 2 (Init) take all monitor_states in account that are not 0 (Not). There is no reason not to include information about any monitor that is not disabled.

unittests for service_type 7 program/output for empty element and for the unlikely case of more than one list elements.